requirements.txt
tqdm
PyYAML
gdown
psycopg2-binary

numpy
matplotlib
pandas
Pillow
opencv-python-headless
ultralytics

streamlit